backend/requirements.txt
webvtt-py
requests
psycopg2-binary
dotenv
Flask
flask-cors
